Changed hover styles
