+++ plugin-hrm-form/src/translations/en-US/flexUI.json
@@ -107,5 +107,9 @@
   "CaseList-ExpandButton": "Open case details",
   "CaseList-PrevPage": "Previous page",
   "CaseList-NextPage": "Next page",
-  "CaseList-SomethingWentWrog": "Something went wrong while loading cases list. Please retry in a moment."
+  "CaseList-SomethingWentWrog": "Something went wrong while loading cases list. Please retry in a moment.",
+
+  "ContactPreview-CopyButton": "Copy informaton",
+  "ContactPreview-ExpandButton": "Open case details",
+  "ContactPreview-MoreOptionsButton": "More options"
 }
